requirements.txt
# requirements.txt
ccxt
pandas
numpy
feedparser
requests
openai
python-dotenv
asyncpraw
